requirements.txt
httpretty == 0.8.10
hypothesis == 1.10.3
nose == 1.3.7
requests == 2.7.0
python_dateutil == 2.4.2
